% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BoxM.F.R
\name{print.BoxM.F}
\alias{print.BoxM.F}
\title{Print method for a BoxM.F object}
\usage{
\method{print}{BoxM.F}(x)
}
\arguments{
\item{x}{an object of class \code{"BoxM.F"}}
}
\description{
Prints an object of class "BoxM.F", , produced by
\code{\link[smstests]{BoxM.F}}
}
\examples{
data(skulls)
resBoxM.F <- BoxM.F(skulls, Period)
print(resBoxM.F)

}
