% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Levenetests2s.mv.R
\name{print.Levene.t}
\alias{print.Levene.t}
\title{Print method for a "Levene.t" object}
\usage{
\method{print}{Levene.t}(x)
}
\arguments{
\item{x}{an object of class \code{"Levene.t"}}
}
\description{
Prints an object of class "Levene.t", , produced by
\code{\link[smstests]{Levenetests2s.mv}}
}
\examples{
data(sparrows)
res.Levene.t <- Levenetests2s.mv(sparrows, Survivorship, "S",
                               alternative = "less", var.equal = TRUE,
                               P.adjust = "bonferroni")
print(res.Levene.t)
}
