% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.LeveneT2.R
\name{print.LeveneT2}
\alias{print.LeveneT2}
\title{Print method for a "LeveneT2" object}
\usage{
\method{print}{LeveneT2}(x)
}
\arguments{
\item{x}{an object of class \code{"LeveneT2"}}
}
\description{
Prints an object of class "LeveneT2"
}
\examples{
data(sparrows)
LeveneT2.sparrows <- LeveneT2(sparrows, group = Survivorship, level1 = "S",
                              var.equal = TRUE)
print(LeveneT2.sparrows)

}
