% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MANOVA.mat.R
\name{print.MANOVA.mat}
\alias{print.MANOVA.mat}
\title{Print method for a "MANOVA.mat" object}
\usage{
\method{print}{MANOVA.mat}(x)
}
\arguments{
\item{x}{an object of class \code{"MANOVA.mat"}}
}
\description{
Prints an object of class "MANOVA.mat", produced by
\code{\link[smstests]{MANOVA.mat}}
}
\examples{
data(skulls)
res.MANOVA <- MANOVA.mat(skulls, group = Period)
# Brief output
print(res.MANOVA)

}
