% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Penrose.dist.R
\name{print.Penrose.dist}
\alias{print.Penrose.dist}
\title{Prints an object of class "Penrose.dist"}
\usage{
\method{print}{Penrose.dist}(x)
}
\arguments{
\item{x}{an object of class \code{"Penrose.dist"}}
}
\description{
Prints an object of class "Penrose.dist"
}
\examples{
data(skulls)
res.Penrose <- Penrose.dist(skulls, Period)
print(res.Penrose)

}
