% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Hotelling.mat.R
\name{print.T2}
\alias{print.T2}
\title{Print method for a "Hotelling.mat" object}
\usage{
\method{print}{T2}(x)
}
\arguments{
\item{x}{an object of class \code{"T2"}}
}
\description{
Prints an object of class "T2", produced by
\code{\link[smstests]{Hotelling.mat}}
}
\examples{
data(sparrows)
results.T2 <- Hotelling.mat(sparrows, group = Survivorship, level1 = "S")
print(results.T2)

}
