% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smstests-package.R
\docType{package}
\name{smstests-package}
\alias{smstests}
\alias{smstests-package}
\title{Simple multivariate tests for one single factor}
\description{
The goal of smstests is to facilitate the execution of simple statistical
tests for multivariate data with respect to one single factor.
}
\author{
\strong{Maintainer}: Jorge Navarro Alberto \email{ganava4@gmail.com} (\href{https://orcid.org/0000-0002-5027-6908}{ORCID})

}
\keyword{internal}
