% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BoxM.F.R
\name{summary.BoxM.F}
\alias{summary.BoxM.F}
\title{Summarize Box'M test based on an F-statistic}
\usage{
\method{summary}{BoxM.F}(x, long = FALSE)
}
\arguments{
\item{x}{an object of class BoxM.F}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default). The long output shows the
covariance matrix for each group and the pooled covariance matrix.}
}
\description{
Summarize the results produced by BoxM.F function, with the option to display
the matrices involved in the calculations
}
\examples{
data(skulls)
resBoxM.F <- BoxM.F(skulls, Period)
# Long output
summary.BoxM.F(resBoxM.F, long = TRUE)

}
