% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Levenetests2s.mv.R
\name{summary.Levene.t}
\alias{summary.Levene.t}
\title{Summary method for "Levene.t"}
\usage{
\method{summary}{Levene.t}(x)
}
\arguments{
\item{x}{an object of class "Levene.t"}
}
\description{
Summarize the results produced by \code{\link[smstests]{Levenetests2s.mv}},
consisting of two-sample Levene's tests computed from two-sample t-tests
applied to absolute differences around medians for more than one response
vector
}
\details{
Summarize
}
\examples{
data(sparrows)
res.Levene.t <- Levenetests2s.mv(sparrows, Survivorship, "S",
                               alternative = "less", var.equal = TRUE,
                               P.adjust = "bonferroni")
summary.Levene.t(res.Levene.t)

}
