% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LeveneT2.R
\name{summary.LeveneT2}
\alias{summary.LeveneT2}
\title{Summarize Levene's test based on Hotelling's \eqn{T^2} test}
\usage{
\method{summary}{LeveneT2}(x, long = FALSE)
}
\arguments{
\item{x}{an object of class \code{"LeveneT2"}}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default)}
}
\description{
Summarizes the results produced by \code{LeveneT2}
}
\examples{
data(sparrows)
LeveneT2.sparrows <- LeveneT2(sparrows, group = Survivorship, level1 = "S",
                              var.equal = TRUE)
# Long output
summary.LeveneT2(LeveneT2.sparrows, long = TRUE)

}
