% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MANOVA.mat.R
\name{summary.MANOVA.mat}
\alias{summary.MANOVA.mat}
\title{Summarize a one-way MANOVA}
\usage{
\method{summary}{MANOVA.mat}(
  x,
  test = c("Pillai", "Wilks", "Hotelling-Lawley", "Roy"),
  long = FALSE
)
}
\arguments{
\item{x}{an object of class MANOVA.mat}

\item{test}{The name of the test statistic to be used (the four tests
implemented in \code{summary.manova}). Pillai's test is the default. Partial
matching is used so the name can be abbreviated.}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default)}
}
\description{
Summarizes the results produced by the \code{MANOVA.mat} function
}
\examples{
data(skulls)
res.MANOVA <- MANOVA.mat(skulls, group = Period)
# Long output, Wilks' test
summary.MANOVA.mat(res.MANOVA, test = "Wilks", long = TRUE)

}
