% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Penrose.dist.R
\name{summary.Penrose.dist}
\alias{summary.Penrose.dist}
\title{Summarize calculations of the Penrose distance}
\usage{
\method{summary}{Penrose.dist}(x, long = FALSE)
}
\arguments{
\item{x}{an object of class Penrose.dist}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default). In addition to Penrose's
distances, the long output displays the covariance matrix for each group with
their sample sizes, the mean vector for each group, and the pooled covariance
matrix.}
}
\description{
Summarize the results produced by Penrose.dist
}
\examples{
data(skulls)
res.Penrose <- Penrose.dist(x = skulls, group = Period)
# Short output
summary.Penrose.dist(res.Penrose)

}
