% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Hotelling.mat.R
\name{summary.T2}
\alias{summary.T2}
\title{Summarizing Hotelling's \eqn{T^2} test}
\usage{
\method{summary}{T2}(x, long = FALSE)
}
\arguments{
\item{x}{an object of class \code{"T2"}}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default)}
}
\description{
Summarizes the results produced by the \code{Hotelling.mat} function
}
\examples{
data(sparrows)
results.T2 <- Hotelling.mat(sparrows, group = Survivorship, level1 = "S")
# Long output
summary.T2(results.T2, long = TRUE)

}
