% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ttests2s.mv.R
\name{summary.ttests}
\alias{summary.ttests}
\title{Summarize multiple two-sample t-tests for a multivariate data set}
\usage{
\method{summary}{ttests}(x)
}
\arguments{
\item{x}{an object of class \code{"ttests"}}
}
\description{
Summarizes the results produced by \code{\link[smstests]{ttests2s.mv}}
}
\examples{
data(sparrows)
ttests.sparrows <- ttests2s.mv(sparrows, group = Survivorship, level1 = "S",
                              var.equal = TRUE, P.adjust = "holm")
summary.ttests(ttests.sparrows)

}
