% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttests2s.mv.R
\name{ttests2s.mv}
\alias{ttests2s.mv}
\title{Multiple two-sample t-tests for multivariate data}
\usage{
ttests2s.mv(
  x,
  group,
  level1,
  alternative = "two.sided",
  var.equal = FALSE,
  P.adjust = "none"
)
}
\arguments{
\item{x}{A data frame with one two-level factor and \emph{p} response variables.}

\item{group}{Two-level factor defining groups. It must be one of the columns
in \code{x}.}

\item{level1}{A character string identifying Sample 1. The string must be one
of the factor levels in \code{group}.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}. You can specify just the initial letter.}

\item{var.equal}{a logical variable indicating whether to treat the two
variances as being equal. If \code{TRUE} then the pooled variance is used to
estimate the variance otherwise the Welch (or Satterthwaite) approximation to
the degrees of freedom is used.}

\item{P.adjust}{p-value correction method, a character string. Can be
abbreviated.}
}
\value{
Returns an object of class \code{"ttests"}, a list containing the
following components:
\tabular{ll}{
\code{name} \tab A character string describing the function \cr
\code{t.list} \tab A list containing \emph{p} vectors of length 3, each vector
having the computed t-statistic, the degrees of freedom for the
t-statistic and the adjusted p-value for the test, respectively.
}

The extractor function \code{\link[smstests]{summary.ttests}} returns an
annotated output of the t-tests.
}
\description{
Performs multiple two-sample t-tests on more than one response vector with
corrected significance levels using any of the adjustment methods for
multiple comparisons offered by \code{p.adjust}.
}
\details{
This function extends the univariate \code{t.test} for the comparison of mean
values for two samples, when more than one variable is involved in the data
analysis, so that type one error rates ("false significances") in a series of
univariate t-tests are adjusted according to the number of response
variables analyzed. The pairwise comparisons between the two levels in
\code{group} with corrections for multiple testing are made over more than
one response vector thus, the function is a variation of
\code{\link[stats]{pairwise.t.test}}.

The methods implemented are the same as those contained in the
\code{p.adjust.methods} for \code{p.adjust}: \code{"bonferroni"},
\code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{"BH"}
(Benjamini-Hochberg) or its alias \code{"fdr"} (False Discovery Rate), and
\code{"BY"} (Benjamini & Yekutieli). The default pass-through option
(\code{"none"}) is also included.
}
\examples{
data(sparrows)
ttests.sparrows <- ttests2s.mv(sparrows, group = Survivorship, level1 = "S",
                              var.equal = TRUE, P.adjust = "bonferroni")
summary.ttests(ttests.sparrows)

}
\author{
Jorge Navarro Alberto, \email{ganava4@gmail.com}
}
